implementacion ciclo ¿ osciloscopio o generado ?
se implemento el ciclo que va a determinar si este es osciloscopio
o es un generador se agrego una sentencial a la funcion osciloscopio
y se implemento la funcion en el archivo de python , se empezo a
adecuar la funcion de la onda senoidal.

+++ Miclase.h
@@ -12,7 +12,7 @@ private:
     //char canal1;
     //char canal2;
     char option;
-    int valor_sensor;
+    float valor_sensor;
     float voltaje;
 };
